requirements.txt
numpy==1.17.1
matplotlib==3.1.1
Keras==2.2.5
tensorflow==1.14.0
Pillow==6.1.0
